refactored to run on game tick
