--- src/main/java/com/runedeck/LogoutMessage.java
@@ -1,7 +1,0 @@
-package com.runedeck;
-
-public class LogoutMessage extends StreamDeckMessage{
-    public LogoutMessage() {
-        super("logout");
-    }
-}
