+++ src/main/java/com/runedeck/LogoutPayload.java
@@ -1,0 +1,7 @@
+package com.runedeck;
+
+public class LogoutPayload extends StreamDeckPayload {
+    public LogoutPayload() {
+        super("logout");
+    }
+}
