--- src/main/java/com/runedeck/StreamDeckMessage.java
@@ -1,13 +1,0 @@
-package com.runedeck;
-
-abstract public class StreamDeckMessage {
-    private final String messageType;
-
-    protected StreamDeckMessage(String messageType) {
-        this.messageType = messageType;
-    }
-
-    public String getMessageType() {
-        return messageType;
-    }
-}
