+++ src/main/java/com/runedeck/WorldsPayload.java
@@ -1,0 +1,16 @@
+package com.runedeck;
+
+import net.runelite.api.Client;
+import net.runelite.api.World;
+
+public class WorldsPayload extends StreamDeckPayload {
+    private final World[] worlds;
+    public WorldsPayload(Client client){
+        super("worlds");
+        this.worlds = client.getWorldList();
+    }
+
+    public World[] getWorlds() {
+        return worlds;
+    }
+}
